+++ src/monugstore/__init__.py
@@ -1,4 +1,4 @@
+from .credentials import ServiceAccountCredentials
 from .manager import GCSManager
-from .oauth import OauthHandler
 
-__all__ = ["GCSManager", "OauthHandler"]
+__all__ = ["GCSManager", "ServiceAccountCredentials"]